fix for sunrise/sunset automation
