+++ config.py
@@ -101,41 +101,3 @@ CITIES = [
     # Hawaii Time Zone
     {"name": "Honolulu, HI", "lat": 21.3099, "lon": -157.8581, "timezone": "Pacific/Honolulu"},
 ]
-
-def calculate_sun_times(city_name, date=None):
-    """Calculate sunrise and sunset times for a given city and date"""
-    if date is None:
-        date = datetime.date.today()
-    
-    # Find the city in our database
-    city_data = None
-    for city in CITIES:
-        if city["name"] == city_name:
-            city_data = city
-            break
-    
-    if not city_data:
-        return None, None
-    
-    try:
-        # Create location info for astral calculations
-        location = LocationInfo(
-            name=city_data["name"],
-            region="USA",
-            latitude=city_data["lat"],
-            longitude=city_data["lon"],
-            timezone=city_data["timezone"]
-        )
-        
-        # Calculate sun times
-        s = sun(location.observer, date=date, tzinfo=pytz.timezone(city_data["timezone"]))
-        
-        # Extract sunrise and sunset times
-        sunrise = s["sunrise"].time()
-        sunset = s["sunset"].time()
-        
-        return sunrise, sunset
-        
-    except Exception as e:
-        print(f"Error calculating sun times for {city_name}: {e}")
-        return None, None
