+++ settings.py
@@ -19,6 +19,9 @@ import functools
 import socket
 from scheduler import weather_update_lock, update_weather as scheduler_update_weather
 from led_test import test_leds, turn_off_leds, update_brightness
+import pytz
+from astral import LocationInfo
+from astral.sun import sun
 
 def after_this_response(func):
     @functools.wraps(func)
@@ -281,7 +284,6 @@ def edit_settings():
                 # If using sunrise/sunset, calculate and update the times
                 if 'use_sunrise_sunset' in request.form:
                     try:
-                        from config import calculate_sun_times
                         sunrise, sunset = calculate_sun_times(request.form['selected_city'])
                         if sunrise and sunset:
                             config_updates["BRIGHT_TIME_START"] = f"datetime.time({sunrise.hour}, {sunrise.minute})"
@@ -1654,7 +1656,6 @@ def calculate_sun_times_endpoint():
         if not city_name:
             return jsonify({'error': 'City name is required'}), 400
             
-        from config import calculate_sun_times
         sunrise, sunset = calculate_sun_times(city_name)
         
         if sunrise and sunset:
@@ -1680,6 +1681,44 @@ def get_cities():
         return jsonify({'cities': cities})
     except Exception as e:
         return jsonify({'error': f'Error getting cities: {str(e)}'}), 500
+
+def calculate_sun_times(city_name, date=None):
+    """Calculate sunrise and sunset times for a given city and date"""
+    if date is None:
+        date = datetime.date.today()
+    
+    # Find the city in our database
+    city_data = None
+    for city in config.CITIES:
+        if city["name"] == city_name:
+            city_data = city
+            break
+    
+    if not city_data:
+        return None, None
+    
+    try:
+        # Create location info for astral calculations
+        location = LocationInfo(
+            name=city_data["name"],
+            region="USA",
+            latitude=city_data["lat"],
+            longitude=city_data["lon"],
+            timezone=city_data["timezone"]
+        )
+        
+        # Calculate sun times
+        s = sun(location.observer, date=date, tzinfo=pytz.timezone(city_data["timezone"]))
+        
+        # Extract sunrise and sunset times
+        sunrise = s["sunrise"].time()
+        sunset = s["sunset"].time()
+        
+        return sunrise, sunset
+        
+    except Exception as e:
+        print(f"Error calculating sun times for {city_name}: {e}")
+        return None, None
 
 if __name__ == '__main__':
     if ENABLE_HTTPS:
